requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
frozenlist==1.8.0
idna==3.10
lxml==6.0.2
magic-filter==1.0.12
multidict==6.7.0
numpy==2.3.3
pandas==2.3.3
propcache==0.4.1
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
yarl==1.22.0
